evaluations/cooking-agent/requirements.txt
# Cooking AI Agent Requirements
# Note: The --pre flag is REQUIRED while Agent Framework is in preview

# Microsoft Agent Framework with Azure AI integration
# Install with: pip install -r requirements.txt --pre
agent-framework-azure-ai

# OpenAI SDK (for Azure OpenAI)
openai>=1.0.0

# Azure authentication
azure-identity>=1.14.0

# Additional utilities
python-dotenv>=1.0.0
